Created speed control (1-10)
Created a speed control for all four turning and driving functions. Those functions now have another parameter (by default 10 (max)) - speedfactor, which takes an int between 0 and 10. Because it's set by default, all old sketches should work. It has not been tested in Arduino. An problem that occurs is that the distance time to drive or turn for a certain distance/amount of degrees is not accurate anymore when speed_factor is not 10.